+++ CHANGELOG.md
@@ -5,10 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]
+[2.3.0] 2026-03-05
+### Add
+- Add `.lang-src` support do render content in source language hidden but expandable.
 ### Changed
+- Changed how `[fndef]` renders footnote's content
 - add border bottom to single's post-header
 - refine reading-time padding
+### Fixed
+- Multi-paragraph footnote's content is now handled correctly by footnotes's popups presenting each paragraph on a new page.
 
 [2.2.0] 2026-03-02
 ### Changed
@@ -364,7 +369,8 @@ the slug's name
 - Initial `index.php`, `style.css`, and `functions.php`
 - This Changelog
 
-[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v2.2.0...HEAD
+[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v2.3.0...HEAD
+[2.3.0]: https://github.com/erremauro/cigno-zen/releases/tag/v2.3.0
 [2.2.0]: https://github.com/erremauro/cigno-zen/releases/tag/v2.2.0
 [2.1.0]: https://github.com/erremauro/cigno-zen/releases/tag/v2.1.0
 [2.0.1]: https://github.com/erremauro/cigno-zen/releases/tag/v2.0.1